requirements.txt
paho-mqtt>=1.6.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
websockets>=12.0
python-multipart>=0.0.6
pytz>=2024.1
python-dotenv>=1.0.0
apscheduler>=3.10.4
